finalizada parcialmente validações da api

+++ src/main/java/com/deliveryfood/domain/service/CadastroCozinhaService.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.deliveryfood.domain.exception.EntidadeEmUsoException;
@@ -16,33 +15,32 @@ import com.deliveryfood.domain.repository.CozinhaRepository;
 
 @Service
 public class CadastroCozinhaService {
+	private static final String COZINHA_COM_O_CÓDIGO_D_ESTÁ_EM_USO = "Cozinha com o código %d está em uso";
+	private static final String COZINHA_COM_O_CÓDIGO_D_NÃO_ENCONTRADA = "Cozinha com o código %d não encontrada";
+
 	@Autowired
 	private CozinhaRepository cozinhaRepository;
 
 	public Cozinha salvar(Cozinha cozinha) {
-
 		return cozinhaRepository.save(cozinha);
 	}
 
 	public Cozinha alterar(Long id, Cozinha cozinha) {
-		Optional<Cozinha> optionalCozinha = cozinhaRepository.findById(id);
-
-		Cozinha cozinhaAtual = optionalCozinha.orElseThrow(
-				() -> new EntidadeNaoEncontradaException(String.format("Cozinha com o código %d não encontrada", id)));
-
+		Cozinha cozinhaAtual = this.buscarOutFalhar(id);
 		BeanUtils.copyProperties(cozinha, cozinhaAtual, "id");
 
 		return cozinhaRepository.save(cozinhaAtual);
-
 	}
 
 	public void excluir(Long id) {
+		if (!cozinhaRepository.existsById(id)) {
+			throw new EntidadeNaoEncontradaException(String.format(COZINHA_COM_O_CÓDIGO_D_NÃO_ENCONTRADA, id));
+		}
+
 		try {
 			cozinhaRepository.deleteById(id);
-		} catch (EmptyResultDataAccessException e) {
-			throw new EntidadeNaoEncontradaException(String.format("Cozinha com o código %d não encontrada", id));
 		} catch (DataIntegrityViolationException e) {
-			throw new EntidadeEmUsoException(String.format("Cozinha com o código %d está em uso", id));
+			throw new EntidadeEmUsoException(String.format(COZINHA_COM_O_CÓDIGO_D_ESTÁ_EM_USO, id));
 		}
 	}
 
@@ -51,14 +49,18 @@ public class CadastroCozinhaService {
 	}
 
 	public Cozinha findById(Long id) {
-		Optional<Cozinha> optionalCozinha = cozinhaRepository.findById(id);
-		return optionalCozinha.orElseThrow(
-				() -> new EntidadeNaoEncontradaException(String.format("Cozinha com o código %d não encontrada", id)));
-
+		return this.buscarOutFalhar(id);
 	}
 
 	public List<Cozinha> findByNomeContaining(String nome) {
-
 		return cozinhaRepository.findByNomeContaining(nome);
+	}
+
+	public Cozinha buscarOutFalhar(Long id) {
+		Optional<Cozinha> optionalCozinha = cozinhaRepository.findById(id);
+
+		return optionalCozinha.orElseThrow(
+				() -> new EntidadeNaoEncontradaException(String.format(COZINHA_COM_O_CÓDIGO_D_NÃO_ENCONTRADA, id)));
+
 	}
 }
